% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_public_holidays.R
\name{add_holidays}
\alias{add_holidays}
\title{Add information whether a day is a public holiday or not}
\usage{
add_holidays(min_date, max_date, all_times)
}
\arguments{
\item{min_date}{start date}

\item{max_date}{end date}

\item{all_times}{data.frame with a date_time column}
}
\value{
same data.frame with additional column "is_public_holiday" (0/1)
}
\description{
Add information whether a day is a public holiday or not
}
