% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{add_school_holidays}
\alias{add_school_holidays}
\title{Add information whether a day is a school holiday or not}
\usage{
add_school_holidays(min_date, max_date, all_times)
}
\arguments{
\item{min_date}{(POSIXct) start date}

\item{max_date}{(POSIXct) end date}

\item{all_times}{data.frame with a date_time column}
}
\value{
same data.frame with additional column "type_school_holiday"s
}
\description{
Add information whether a day is a school holiday or not
}
