% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{add_year_to_date}
\alias{add_year_to_date}
\title{Add year to date}
\usage{
add_year_to_date(dat, year)
}
\arguments{
\item{dat}{(data.frame) with dates}

\item{year}{(int)}
}
\value{
data.frame completed with given year
}
\description{
This function adds the given year to a given date which
only consist out of month and day
}
