% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_complement_times.R
\name{create_hourly_timestamps}
\alias{create_hourly_timestamps}
\title{Create hourly timestamps between a given date end date}
\usage{
create_hourly_timestamps(min_date, max_date)
}
\arguments{
\item{min_date}{start date}

\item{max_date}{end date}
}
\value{
data.frame with a date_time column
}
\description{
Create hourly timestamps between a given date end date
}
