% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{create_raw_school_holidays}
\alias{create_raw_school_holidays}
\title{Create a data frame with dates of school holidays in a given year}
\usage{
create_raw_school_holidays(year)
}
\arguments{
\item{year}{(int) the year that should be crawled}
}
\value{
data.frame with start and end date of the holidays in given year
}
\description{
data frame consists of type of holiday, start date of the holiday,
end date of the holiday. Data is crawled from www.schulferien.org
}
