% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{create_school_holidays}
\alias{create_school_holidays}
\title{Creates a dataframe with school holidays}
\usage{
create_school_holidays(min_date, max_date)
}
\arguments{
\item{min_date}{(POSIXct) start date}

\item{max_date}{(POSIXct) end date}
}
\value{
data.frame with school holidays
}
\description{
dataframe consists of equidistant hourly spaced "date_time" column
and "type_school_holiday" where the name of the holiday is stored.
}
