% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_public_holidays.R
\name{download_holidays}
\alias{download_holidays}
\title{Download holidays for years 2015 to 2024}
\usage{
download_holidays(years)
}
\arguments{
\item{years}{numeric vector with years for which we retrieve holidays}
}
\value{
data.frame with columns date and "is_public_holiday" (always 1)
}
\description{
Download holidays for years 2015 to 2024
}
