% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{filter_long_weekend}
\alias{filter_long_weekend}
\title{Filters out elements of data frame with "long weekends"}
\usage{
filter_long_weekend(dat)
}
\arguments{
\item{dat}{(data.frame) with school holiday with combined "long weekend"}
}
\value{
processed data.frame with separate "long weekends"
}
\description{
Some dates of school holidays are not given by an
closed interval. They are given by a string with separators. The char "+"
separates single days, the char "-" indicates an interval.
e.g. "09.07.+10.07.-22.08."
This function separates these and stores them in multiple rows of the given data.frame
}
