% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{get_data_schulferien_org}
\alias{get_data_schulferien_org}
\title{Web crawls schulferien.org}
\usage{
get_data_schulferien_org(year)
}
\arguments{
\item{year}{(int) year to web crawl data}
}
\value{
data.frame with all school holidays of given year
}
\description{
crawls all German school holidays from schulferien.org of given year
}
