% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{get_years}
\alias{get_years}
\title{get years from date}
\usage{
get_years(min_date, max_date)
}
\arguments{
\item{min_date}{(POSIXct) Start Date}

\item{max_date}{(POSIXct) End Date}
}
\value{
int vector containing all years
}
\description{
This function calculates a vector with the years from a start and end date
The start year is the year before the start date to make sure that Christmas
holidays are completely included.
}
