% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{interval_to_hourly_sequence}
\alias{interval_to_hourly_sequence}
\title{Convert a data frame with start and end date to a data frame with an hourly
spaced time interval}
\usage{
interval_to_hourly_sequence(df)
}
\arguments{
\item{df}{(data.frame) with 3 Columns.
1st Column - type_school_holiday,
2nd Column - date_start,
3rd Column - date_end}
}
\value{
df_new (data.frame) with 2 Columns
1st Column - Event Name,
2nd Column date_time with an hourly time stamp
}
\description{
This function transforms a given data frame with the holiday
name, start date and end date a data frame with same event name but with an
hourly spaced time interval.
}
