% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{preprocess_data_schulferien}
\alias{preprocess_data_schulferien}
\title{Prepares Data from schulferien.org}
\usage{
preprocess_data_schulferien(dat)
}
\arguments{
\item{dat}{(data.frame) data.frame given by schulferien.org}
}
\value{
processed data.frame
}
\description{
This function selects only the data of the federal State Berlin.
Converts the data frame from wide to long, removes all white spaces, asterisks and
drops all empty rows of data frame.
}
