% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_school_holidays.R
\name{separate_date_interval}
\alias{separate_date_interval}
\title{Stores beginning and end of holidays in 2 separate variables}
\usage{
separate_date_interval(dat)
}
\arguments{
\item{dat}{(data.frame) with combined date interval}
}
\value{
data.frame with separated start and end date
}
\description{
This function separates an date interval in start
and end and replaces them in the given data.frame
}
